showe products by date of day

+++ assets/Products.js
@@ -59,31 +59,34 @@ class Products {
     }
 
     renderProductsTable() {
+        const today = new Date();
+        const formattedDate = today.toISOString().split('T')[0];
         const tableBody = document.getElementById("productTable").querySelector("tbody");
         tableBody.innerHTML = "";
 
         for (let i = 0; i < this.productsArray.length; i++) {
-            const productRow = document.createElement("tr");
-            productRow.innerHTML = `
-                <td>${this.productsArray[i].code}</td>
-                <td>${this.productsArray[i].name}</td>
-                <td>${this.productsArray[i].price}</td>
-                <td>${this.productsArray[i].category}</td>
-                <td class="d-none d-sm-block">${this.productsArray[i].description}</td>
-                <td>
-                    <div
-                        class="d-flex flex-row justify-content-between gap-2 align-items-center px-1">
-                        <i class="fa-solid fa-pen-to-square d-block update-btn" role="button"></i>
-                        <i class="fa-solid fa-trash-can d-block delete-btn" role="button"></i>
-                    </div>
-                </td>
-            `;
-            tableBody.appendChild(productRow);
-
-            productRow.querySelector(".delete-btn").addEventListener("click", () => {
-                this.deleteProduct(i);
-            });
-
+            if( this.productsArray[i].date.includes(formattedDate)){
+                const productRow = document.createElement("tr");
+                productRow.innerHTML = `
+                    <td>${this.productsArray[i].code}</td>
+                    <td>${this.productsArray[i].name}</td>
+                    <td>${this.productsArray[i].price}</td>
+                    <td>${this.productsArray[i].category}</td>
+                    <td class="d-none d-sm-block">${this.productsArray[i].description}</td>
+                    <td>
+                        <div
+                            class="d-flex flex-row justify-content-between gap-2 align-items-center px-1">
+                            <i class="fa-solid fa-pen-to-square d-block update-btn" role="button"></i>
+                            <i class="fa-solid fa-trash-can d-block delete-btn" role="button"></i>
+                        </div>
+                    </td>
+                `;
+                tableBody.appendChild(productRow);
+    
+                productRow.querySelector(".delete-btn").addEventListener("click", () => {
+                    this.deleteProduct(i);
+                });
+            }
         }
     }
